refactor(lib): move lib.rs logic into dedicated modules
Extract all logic out of five crate roots so each lib.rs contains only
crate attributes, module declarations, and re-exports (audit item A1):

- base-execution-cli: Cli -> cli.rs
- ingress-rpc-lib: Config -> config.rs, BuilderConnector -> builder.rs
- base-proof-succinct-proof-utils: cluster utilities -> cluster.rs
- base-execution-consensus: BaseBeaconConsensus -> beacon.rs
- base-execution-exex: BaseProofsExEx -> exex.rs

Also move the [features] section below [dev-dependencies] in
base-consensus-providers and base-consensus-sources (audit item A8).

Extractions preserve behavior verbatim; workspace builds, clippy is
clean, and nightly fmt applied.
