+++ crates/execution/exex/src/exex.rs
@@ -1,0 +1,1133 @@
+//! Base proofs `ExEx`: builder and execution-extension driver.
+
+use std::{sync::Arc, time::Duration};
+
+use alloy_consensus::BlockHeader;
+use alloy_eips::eip1898::BlockWithParent;
+#[cfg(feature = "metrics")]
+use base_execution_trie::BaseProofsStore;
+use base_execution_trie::{
+    BaseProofStoragePrunerTask, BaseProofsBatchStore, BaseProofsStorage,
+    live::{BatchBlock, LiveTrieCollector},
+    metrics::BlockMetrics,
+};
+use futures::TryStreamExt;
+use reth_execution_types::Chain;
+use reth_exex::{ExExContext, ExExEvent, ExExNotification, ExExNotificationsStream};
+use reth_node_api::{FullNodeComponents, NodePrimitives, NodeTypes};
+use reth_provider::{BlockNumReader, BlockReader, TransactionVariant};
+use tokio::task;
+use tracing::{debug, error, info};
+
+use crate::sync_target::{CachedBlockTrieData, SyncTarget, SyncTargetState};
+
+/// Default safety threshold for the gap between stored earliest block and the configured
+/// window target. When exceeded on startup, the node refuses to auto-prune and asks the
+/// operator to run `proofs prune` manually. The threshold is a backstop against accidental
+/// misconfiguration (e.g. shrinking the proofs window by orders of magnitude) — it is not a
+/// measure of pruner throughput. Override via
+/// [`BaseProofsExExBuilder::with_max_prune_blocks_startup`].
+const DEFAULT_MAX_PRUNE_BLOCKS_STARTUP: u64 = 100_000;
+
+/// How many blocks to process in a single sync turn before yielding.
+const SYNC_BLOCKS_PER_TURN: usize = 50;
+
+/// Default proofs history window: 1 month of blocks at 2s block time
+const DEFAULT_PROOFS_HISTORY_WINDOW: u64 = 1_296_000;
+
+/// Default interval between proof-storage prune runs. Default is 15 seconds.
+const DEFAULT_PRUNE_INTERVAL: Duration = Duration::from_secs(15);
+
+/// Default verification interval: disabled
+const DEFAULT_VERIFICATION_INTERVAL: u64 = 0; // disabled
+
+/// Builder for [`BaseProofsExEx`].
+#[derive(Debug)]
+pub struct BaseProofsExExBuilder<Node, Storage>
+where
+    Node: FullNodeComponents,
+{
+    ctx: ExExContext<Node>,
+    storage: BaseProofsStorage<Storage>,
+    proofs_history_window: u64,
+    proofs_history_prune_interval: Duration,
+    verification_interval: u64,
+    max_prune_blocks_startup: u64,
+}
+
+impl<Node, Storage> BaseProofsExExBuilder<Node, Storage>
+where
+    Node: FullNodeComponents,
+{
+    /// Create a new builder with required parameters and defaults.
+    pub const fn new(ctx: ExExContext<Node>, storage: BaseProofsStorage<Storage>) -> Self {
+        Self {
+            ctx,
+            storage,
+            proofs_history_window: DEFAULT_PROOFS_HISTORY_WINDOW,
+            proofs_history_prune_interval: DEFAULT_PRUNE_INTERVAL,
+            verification_interval: DEFAULT_VERIFICATION_INTERVAL,
+            max_prune_blocks_startup: DEFAULT_MAX_PRUNE_BLOCKS_STARTUP,
+        }
+    }
+
+    /// Sets the window to span blocks for proofs history.
+    pub const fn with_proofs_history_window(mut self, window: u64) -> Self {
+        self.proofs_history_window = window;
+        self
+    }
+
+    /// Sets the interval between proof-storage prune runs.
+    pub const fn with_proofs_history_prune_interval(mut self, interval: Duration) -> Self {
+        self.proofs_history_prune_interval = interval;
+        self
+    }
+
+    /// Sets the verification interval.
+    pub const fn with_verification_interval(mut self, interval: u64) -> Self {
+        self.verification_interval = interval;
+        self
+    }
+
+    /// Sets the safety threshold for blocks that may be auto-pruned at startup.
+    /// See [`DEFAULT_MAX_PRUNE_BLOCKS_STARTUP`].
+    pub const fn with_max_prune_blocks_startup(mut self, max_blocks: u64) -> Self {
+        self.max_prune_blocks_startup = max_blocks;
+        self
+    }
+
+    /// Builds the [`BaseProofsExEx`].
+    pub fn build(self) -> BaseProofsExEx<Node, Storage> {
+        BaseProofsExEx {
+            ctx: self.ctx,
+            storage: self.storage,
+            proofs_history_window: self.proofs_history_window,
+            proofs_history_prune_interval: self.proofs_history_prune_interval,
+            verification_interval: self.verification_interval,
+            max_prune_blocks_startup: self.max_prune_blocks_startup,
+        }
+    }
+}
+
+/// Proofs `ExEx` - processes blocks and tracks state changes within fault proof window.
+///
+/// Saves and serves trie nodes to make proofs faster. This handles the process of
+/// saving the current state, new blocks as they're added, and serving proof RPCs
+/// based on the saved data.
+///
+/// # Examples
+///
+/// The following example shows how to install the `ExEx` with either in-memory or persistent storage.
+/// This can be used when launching a Base node via a binary.
+///
+/// ```
+/// use futures::FutureExt;
+/// use reth_db::test_utils::create_test_rw_db;
+/// use reth_node_api::NodeTypesWithDBAdapter;
+/// use reth_node_builder::{NodeBuilder, NodeConfig};
+/// use base_execution_chainspec::BaseChainSpec;
+/// use base_execution_exex::BaseProofsExEx;
+/// use base_node_core::{BaseNode, args::RollupArgs};
+/// use base_execution_trie::{InMemoryProofsStorage, BaseProofsStorage, RocksdbProofsStorage};
+/// use reth_provider::providers::BlockchainProvider;
+/// use std::{sync::Arc, time::Duration};
+///
+/// let config = NodeConfig::new(Arc::new(BaseChainSpec::mainnet()));
+/// let db = create_test_rw_db();
+/// let args = RollupArgs::default();
+/// let base_node = BaseNode::new(args);
+///
+/// // Create in-memory or persistent storage
+/// let storage: BaseProofsStorage<Arc<InMemoryProofsStorage>> =
+///     Arc::new(InMemoryProofsStorage::new()).into();
+///
+/// // Example for creating persistent storage
+/// # let temp_dir = tempfile::tempdir().expect("Failed to create temp dir");
+/// # let storage_path = temp_dir.path().join("proofs_storage");
+///
+/// # let storage: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::new(
+/// #    RocksdbProofsStorage::new(&storage_path).expect("Failed to create RocksdbProofsStorage"),
+/// # ).into();
+///
+/// let storage_exec = storage.clone();
+/// let proofs_history_window = 1_296_000u64;
+/// let proofs_history_prune_interval = Duration::from_secs(3600);
+///
+/// // Verification interval: perform full execution every N blocks
+/// let verification_interval = 0; // 0 = disabled, 100 = verify every 100 blocks
+///
+/// // Can also use install_exex_if along with a boolean flag
+/// // Set this based on your configuration or CLI args
+/// let _builder = NodeBuilder::new(config)
+///     .with_database(db)
+///     .with_types_and_provider::<BaseNode, BlockchainProvider<NodeTypesWithDBAdapter<BaseNode, _>>>()
+///     .with_components(base_node.components())
+///     .install_exex("proofs-history", move |exex_context| async move {
+///         Ok(BaseProofsExEx::builder(exex_context, storage_exec)
+///             .with_proofs_history_window(proofs_history_window)
+///             .with_proofs_history_prune_interval(proofs_history_prune_interval)
+///             .with_verification_interval(verification_interval)
+///             .build()
+///             .run()
+///             .boxed())
+///     })
+///     .on_node_started(|_full_node| Ok(()))
+///     .check_launch();
+/// ```
+#[derive(Debug)]
+pub struct BaseProofsExEx<Node, Storage>
+where
+    Node: FullNodeComponents,
+{
+    /// The `ExEx` context containing the node related utilities e.g. provider, notifications,
+    /// events.
+    ctx: ExExContext<Node>,
+    /// The type of storage DB.
+    storage: BaseProofsStorage<Storage>,
+    /// The window to span blocks for proofs history. Value is the number of blocks, received as
+    /// cli arg.
+    proofs_history_window: u64,
+    /// Interval between proof-storage prune runs
+    proofs_history_prune_interval: Duration,
+    /// Verification interval: perform full block execution every N blocks for data integrity.
+    /// If 0, verification is disabled (always use fast path when available).
+    /// If 1, verification is always enabled (always execute blocks).
+    verification_interval: u64,
+    /// Maximum blocks the startup check is willing to schedule for auto-prune. See
+    /// [`DEFAULT_MAX_PRUNE_BLOCKS_STARTUP`].
+    max_prune_blocks_startup: u64,
+}
+
+impl<Node, Storage> BaseProofsExEx<Node, Storage>
+where
+    Node: FullNodeComponents,
+{
+    /// Create a new `BaseProofsExEx` instance.
+    pub fn new(ctx: ExExContext<Node>, storage: BaseProofsStorage<Storage>) -> Self {
+        BaseProofsExExBuilder::new(ctx, storage).build()
+    }
+
+    /// Create a new builder for `BaseProofsExEx`.
+    pub const fn builder(
+        ctx: ExExContext<Node>,
+        storage: BaseProofsStorage<Storage>,
+    ) -> BaseProofsExExBuilder<Node, Storage> {
+        BaseProofsExExBuilder::new(ctx, storage)
+    }
+}
+
+impl<Node, Storage, Primitives> BaseProofsExEx<Node, Storage>
+where
+    Node: FullNodeComponents<Types: NodeTypes<Primitives = Primitives>>,
+    Primitives: NodePrimitives,
+    Storage: BaseProofsBatchStore + Clone + 'static,
+{
+    /// Main execution loop for the `ExEx`
+    pub async fn run(mut self) -> eyre::Result<()> {
+        self.ensure_initialized()?;
+        let sync_target = self.spawn_sync_task();
+
+        // If storage is behind tip, start syncing immediately rather than waiting
+        // for the first notification.
+        let best_block = self.ctx.provider().best_block_number()?;
+        let latest_stored = self.storage.get_latest_block_number()?.map(|(n, _)| n).unwrap_or(0);
+        if latest_stored < best_block {
+            info!(
+                target: "base::exex",
+                latest_stored,
+                best_block,
+                "Storage behind tip, starting sync immediately"
+            );
+            sync_target.update_state(SyncTargetState::SyncUpTo { to: best_block });
+        }
+
+        let prune_task = BaseProofStoragePrunerTask::new(
+            self.storage.clone(),
+            self.ctx.provider().clone(),
+            self.proofs_history_window,
+            self.proofs_history_prune_interval,
+        );
+        self.ctx
+            .task_executor()
+            .spawn_with_graceful_shutdown_signal(|signal| Box::pin(prune_task.run(signal)));
+
+        self.ctx.notifications.set_without_head();
+
+        while let Some(notification) = self.ctx.notifications.try_next().await? {
+            self.handle_notification(notification, &sync_target)?;
+        }
+
+        Ok(())
+    }
+
+    /// Ensure proofs storage is initialized
+    fn ensure_initialized(&self) -> eyre::Result<()> {
+        // Check if proofs storage is initialized
+        let earliest_block_number = match self.storage.get_earliest_block_number()? {
+            Some((n, _)) => n,
+            None => {
+                return Err(eyre::eyre!(
+                    "Proofs storage not initialized. Please run 'base-reth-node proofs init --proofs-history.storage-path <PATH>' first."
+                ));
+            }
+        };
+
+        let latest_block_number = match self.storage.get_latest_block_number()? {
+            Some((n, _)) => n,
+            None => {
+                return Err(eyre::eyre!(
+                    "Proofs storage not initialized. Please run 'base-reth-node proofs init --proofs-history.storage-path <PATH>' first."
+                ));
+            }
+        };
+
+        // Check if we have accumulated too much history for the configured window.
+        // If the gap between what we have and what we want to keep is too large, the auto-pruner
+        // will stall the node.
+        let target_earliest = latest_block_number.saturating_sub(self.proofs_history_window);
+        if target_earliest > earliest_block_number {
+            let blocks_to_prune = target_earliest - earliest_block_number;
+            if blocks_to_prune > self.max_prune_blocks_startup {
+                return Err(eyre::eyre!(
+                    "Configuration requires pruning {} blocks, which exceeds the safety threshold of {}. \
+                     Huge prune operations can stall the node. \
+                     Please run 'base-reth-node proofs prune' manually before starting the node.",
+                    blocks_to_prune,
+                    self.max_prune_blocks_startup
+                ));
+            }
+        }
+
+        // Need to update the earliest block metric on startup as this is not called frequently and
+        // can show outdated info. When metrics are disabled, this is a no-op.
+        BlockMetrics::earliest_number().set(earliest_block_number as f64);
+
+        Ok(())
+    }
+
+    /// Spawn the background sync task and return the shared [`SyncTarget`].
+    ///
+    /// The sync target buffers trie data from notifications so the sync loop
+    /// can use pre-computed data for blocks even when it is many blocks behind
+    /// the chain tip. Blocks whose trie data was evicted from the cache fall
+    /// back to full execution.
+    fn spawn_sync_task(&self) -> Arc<SyncTarget> {
+        let sync_target = Arc::new(SyncTarget::new());
+        let task_sync_target = Arc::clone(&sync_target);
+
+        let task_storage = self.storage.clone();
+        let task_provider = self.ctx.provider().clone();
+        let task_evm_config = self.ctx.evm_config().clone();
+        let verification_interval = self.verification_interval;
+
+        self.ctx.task_executor().spawn_critical_task(
+            "base::exex::proofs_storage_sync_loop",
+            async move {
+                let storage = task_storage.clone();
+                let task_collector =
+                    LiveTrieCollector::new(task_evm_config, task_provider.clone(), &storage);
+                Self::sync_loop(
+                    task_sync_target,
+                    task_storage,
+                    task_provider,
+                    &task_collector,
+                    verification_interval,
+                )
+                .await;
+            },
+        );
+
+        sync_target
+    }
+
+    async fn sync_loop(
+        sync_target: Arc<SyncTarget>,
+        storage: BaseProofsStorage<Storage>,
+        provider: Node::Provider,
+        collector: &LiveTrieCollector<'_, Node::Evm, Node::Provider, Storage>,
+        verification_interval: u64,
+    ) {
+        info!(target: "base::exex", "Starting proofs storage sync loop");
+
+        loop {
+            let Some(state) = sync_target.take_state() else {
+                sync_target.notified().await;
+                continue;
+            };
+
+            match state {
+                SyncTargetState::Revert { revert_to } => {
+                    Self::handle_revert(&storage, collector, revert_to);
+                    sync_target.mark_revert_complete(&revert_to);
+                }
+                SyncTargetState::RevertThenSync { revert_to, sync_to } => {
+                    Self::handle_revert(&storage, collector, revert_to);
+                    sync_target.mark_revert_complete(&revert_to);
+                    Self::sync_forward(
+                        &sync_target,
+                        &storage,
+                        &provider,
+                        collector,
+                        verification_interval,
+                        sync_to,
+                    )
+                    .await;
+                }
+                SyncTargetState::SyncUpTo { to } => {
+                    Self::sync_forward(
+                        &sync_target,
+                        &storage,
+                        &provider,
+                        collector,
+                        verification_interval,
+                        to,
+                    )
+                    .await;
+                }
+            }
+        }
+    }
+
+    fn handle_revert(
+        storage: &BaseProofsStorage<Storage>,
+        collector: &LiveTrieCollector<'_, Node::Evm, Node::Provider, Storage>,
+        revert_to: BlockWithParent,
+    ) {
+        let latest = match storage.get_latest_block_number() {
+            Ok(Some((n, _))) => n,
+            Ok(None) => return,
+            Err(e) => {
+                error!(target: "base::exex", error = ?e, "Failed to get latest block during revert");
+                return;
+            }
+        };
+
+        if latest >= revert_to.block.number {
+            info!(
+                target: "base::exex",
+                revert_to = revert_to.block.number,
+                latest,
+                "Reverting proofs storage"
+            );
+            if let Err(e) = collector.unwind_history(revert_to) {
+                error!(target: "base::exex", error = ?e, "Failed to revert proofs storage");
+            }
+        } else {
+            debug!(
+                target: "base::exex",
+                revert_to = revert_to.block.number,
+                latest,
+                "Revert target beyond stored blocks, skipping"
+            );
+        }
+    }
+
+    async fn sync_forward(
+        sync_target: &SyncTarget,
+        storage: &BaseProofsStorage<Storage>,
+        provider: &Node::Provider,
+        collector: &LiveTrieCollector<'_, Node::Evm, Node::Provider, Storage>,
+        verification_interval: u64,
+        target: u64,
+    ) {
+        loop {
+            // Check for higher-priority state (e.g. revert) before processing.
+            if sync_target.has_pending_state() {
+                return;
+            }
+
+            let latest = match storage.get_latest_block_number() {
+                Ok(Some((n, _))) => n,
+                Ok(None) => {
+                    error!(target: "base::exex", "No blocks stored in proofs storage during sync");
+                    return;
+                }
+                Err(e) => {
+                    error!(target: "base::exex", error = ?e, "Failed to get latest block");
+                    return;
+                }
+            };
+
+            if latest >= target {
+                return;
+            }
+
+            let end = latest.saturating_add(SYNC_BLOCKS_PER_TURN as u64).min(target);
+            info!(
+                target: "base::exex",
+                start = latest + 1,
+                end,
+                target,
+                blocks = end - latest,
+                "Processing proofs storage sync turn"
+            );
+
+            let mut batch: Vec<BatchBlock<Primitives>> =
+                Vec::with_capacity((end - latest) as usize);
+            for block_num in (latest + 1)..=end {
+                let cached = sync_target.take(block_num);
+                match Self::build_batch_entry(block_num, cached, provider, verification_interval) {
+                    Ok(entry) => batch.push(entry),
+                    Err(e) => {
+                        error!(target: "base::exex", block_number = block_num, error = ?e, "Preparing block for batch failed");
+                        return;
+                    }
+                }
+            }
+
+            if let Err(e) = collector.execute_and_store_batch(batch) {
+                error!(target: "base::exex", start = latest + 1, end, error = ?e, "Batch processing failed");
+                return;
+            }
+
+            info!(target: "base::exex", latest_stored = latest, target, "Batch processed, yielding");
+            task::yield_now().await;
+        }
+    }
+
+    fn build_batch_entry(
+        block_number: u64,
+        cached: Option<CachedBlockTrieData>,
+        provider: &Node::Provider,
+        verification_interval: u64,
+    ) -> eyre::Result<BatchBlock<Primitives>> {
+        let should_verify =
+            verification_interval > 0 && block_number.is_multiple_of(verification_interval);
+        let has_cached = cached.is_some();
+
+        if let Some(cached) = cached
+            && !should_verify
+        {
+            let sorted = cached.trie_data.get();
+            debug!(
+                target: "base::exex",
+                block_number,
+                "Using pre-computed state from notification"
+            );
+            return Ok(BatchBlock::Cached {
+                block_with_parent: cached.block_with_parent,
+                sorted_trie_updates: Arc::clone(&sorted.sorted.trie_updates),
+                sorted_post_state: Arc::clone(&sorted.sorted.hashed_state),
+            });
+        }
+
+        if has_cached {
+            info!(
+                target: "base::exex",
+                block_number,
+                verification_interval,
+                "Periodic verification: performing full block execution despite cached data"
+            );
+        } else {
+            debug!(
+                target: "base::exex",
+                block_number,
+                "No cached trie data, falling back to full execution"
+            );
+        }
+
+        debug!(
+            target: "base::exex",
+            block_number,
+            "Fetching block from provider for execution",
+        );
+
+        let block = provider
+            .recovered_block(block_number.into(), TransactionVariant::NoHash)?
+            .ok_or_else(|| eyre::eyre!("Missing block {} in provider", block_number))?;
+
+        Ok(BatchBlock::Execute(Box::new(block)))
+    }
+
+    fn handle_notification(
+        &self,
+        notification: ExExNotification<Primitives>,
+        sync_target: &SyncTarget,
+    ) -> eyre::Result<()> {
+        match &notification {
+            ExExNotification::ChainCommitted { new } => {
+                self.handle_chain_committed(Arc::clone(new), sync_target)?
+            }
+            ExExNotification::ChainReorged { old, new } => {
+                self.handle_chain_reorged(Arc::clone(old), Arc::clone(new), sync_target)?
+            }
+            ExExNotification::ChainReverted { old } => {
+                self.handle_chain_reverted(Arc::clone(old), sync_target)?
+            }
+        }
+
+        if let Some(committed_chain) = notification.committed_chain() {
+            let tip = committed_chain.tip().num_hash();
+            debug!(
+                target: "base::exex",
+                block_number = tip.number,
+                block_hash = ?tip.hash,
+                "Sending FinishedHeight event"
+            );
+            self.ctx.events.send(ExExEvent::FinishedHeight(tip))?;
+        }
+
+        Ok(())
+    }
+
+    fn handle_chain_committed(
+        &self,
+        new: Arc<Chain<Primitives>>,
+        sync_target: &SyncTarget,
+    ) -> eyre::Result<()> {
+        debug!(
+            target: "base::exex",
+            block_number = new.tip().number(),
+            block_hash = ?new.tip().hash(),
+            "ChainCommitted notification received",
+        );
+
+        // Cache trie data for all blocks in the chain so the sync loop can
+        // use pre-computed data even when it is many blocks behind.
+        let total_blocks = new.blocks().len();
+        let mut cached_count = 0usize;
+        for (&block_number, block) in new.blocks() {
+            if let Some(trie_data) = new.trie_data_at(block_number) {
+                sync_target.insert(
+                    block_number,
+                    CachedBlockTrieData {
+                        block_with_parent: block.block_with_parent(),
+                        trie_data: trie_data.clone(),
+                    },
+                );
+                cached_count += 1;
+            } else {
+                debug!(
+                    target: "base::exex",
+                    block_number,
+                    "Notification block missing trie data"
+                );
+            }
+        }
+
+        debug!(
+            target: "base::exex",
+            tip = new.tip().number(),
+            total_blocks,
+            cached_count,
+            missing = total_blocks - cached_count,
+            "Cached notification trie data"
+        );
+
+        sync_target.update_state(SyncTargetState::SyncUpTo { to: new.tip().number() });
+        Ok(())
+    }
+
+    fn handle_chain_reorged(
+        &self,
+        old: Arc<Chain<Primitives>>,
+        new: Arc<Chain<Primitives>>,
+        sync_target: &SyncTarget,
+    ) -> eyre::Result<()> {
+        info!(
+            target: "base::exex",
+            old_block_number = old.tip().number(),
+            old_block_hash = ?old.tip().hash(),
+            new_block_number = new.tip().number(),
+            new_block_hash = ?new.tip().hash(),
+            "ChainReorged notification received",
+        );
+
+        if old.fork_block() != new.fork_block() {
+            return Err(eyre::eyre!(
+                "Fork blocks do not match: old fork block {:?}, new fork block {:?}",
+                old.fork_block(),
+                new.fork_block()
+            ));
+        }
+
+        let first_old = old.first().block_with_parent();
+
+        // Invalidate any cached blocks from the old chain.
+        sync_target.clear_from(first_old.block.number);
+
+        // Cache trie data for all blocks in the new chain.
+        for (&block_number, block) in new.blocks() {
+            if let Some(trie_data) = new.trie_data_at(block_number) {
+                sync_target.insert(
+                    block_number,
+                    CachedBlockTrieData {
+                        block_with_parent: block.block_with_parent(),
+                        trie_data: trie_data.clone(),
+                    },
+                );
+            } else {
+                debug!(
+                    target: "base::exex",
+                    block_number,
+                    "Reorged block missing trie data"
+                );
+            }
+        }
+
+        sync_target.update_state(SyncTargetState::RevertThenSync {
+            revert_to: first_old,
+            sync_to: new.tip().number(),
+        });
+
+        Ok(())
+    }
+
+    fn handle_chain_reverted(
+        &self,
+        old: Arc<Chain<Primitives>>,
+        sync_target: &SyncTarget,
+    ) -> eyre::Result<()> {
+        info!(
+            target: "base::exex",
+            old_block_number = old.tip().number(),
+            old_block_hash = ?old.tip().hash(),
+            "ChainReverted notification received",
+        );
+
+        let first_old = old.first().block_with_parent();
+
+        // Invalidate any cached blocks that are being reverted.
+        sync_target.clear_from(first_old.block.number);
+
+        sync_target.update_state(SyncTargetState::Revert { revert_to: first_old });
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::{collections::BTreeMap, default::Default, sync::Arc, time::Duration};
+
+    use alloy_consensus::private::alloy_primitives::B256;
+    use alloy_eips::{BlockNumHash, NumHash, eip1898::BlockWithParent};
+    use base_execution_trie::{
+        BaseProofsStorage, BaseProofsStore, BlockStateDiff, RocksdbProofsStorage,
+    };
+    use reth_db::test_utils::tempdir_path;
+    use reth_ethereum_primitives::{Block, Receipt};
+    use reth_execution_types::{Chain, ExecutionOutcome};
+    use reth_primitives_traits::RecoveredBlock;
+    use reth_trie::{
+        ComputedTrieData, HashedPostStateSorted, LazyTrieData, updates::TrieUpdatesSorted,
+    };
+
+    use super::*;
+
+    // -------------------------------------------------------------------------
+    // Helpers: deterministic blocks and deterministic Chain with precomputed updates
+    // -------------------------------------------------------------------------
+    fn b256(byte: u8) -> B256 {
+        B256::new([byte; 32])
+    }
+
+    // deterministic hash from block number: 0 -> 0x00.., 1 -> 0x01.., etc.
+    fn hash_for_num(num: u64) -> B256 {
+        // if you only care about small test numbers, this is enough:
+        b256(num as u8)
+
+        // If you want to avoid wrapping when num > 255, use something like:
+        // let mut out = [0u8; 32];
+        // out[0..8].copy_from_slice(&num.to_be_bytes());
+        // B256::new(out)
+    }
+
+    fn mk_block(num: u64) -> RecoveredBlock<Block> {
+        let mut b: RecoveredBlock<Block> = Default::default();
+        b.set_block_number(num);
+        b.set_hash(hash_for_num(num));
+        b.set_parent_hash(hash_for_num(num - 1));
+        b
+    }
+
+    fn mk_chain_with_updates(
+        from: u64,
+        to: u64,
+        hash_override: Option<B256>,
+    ) -> Chain<reth_ethereum_primitives::EthPrimitives> {
+        let mut blocks: Vec<RecoveredBlock<Block>> = Vec::new();
+        let mut trie_data = BTreeMap::new();
+
+        for n in from..=to {
+            let mut b = mk_block(n);
+            if let Some(hash) = hash_override {
+                b.set_hash(hash);
+            }
+            blocks.push(b);
+
+            let data = LazyTrieData::ready(ComputedTrieData::new(
+                Arc::new(HashedPostStateSorted::default()),
+                Arc::new(TrieUpdatesSorted::default()),
+            ));
+            trie_data.insert(n, data);
+        }
+
+        let execution_outcome: ExecutionOutcome<Receipt> = ExecutionOutcome {
+            bundle: Default::default(),
+            receipts: Vec::new(),
+            requests: Vec::new(),
+            first_block: from,
+        };
+
+        Chain::new(blocks, execution_outcome, trie_data)
+    }
+
+    /// Store blocks directly into proofs storage (bypasses the sync loop).
+    fn store_blocks<S: BaseProofsStore>(from: u64, to: u64, storage: &BaseProofsStorage<S>) {
+        for n in from..=to {
+            let chain = mk_chain_with_updates(n, n, None);
+            let block = chain.blocks().get(&n).unwrap();
+            storage
+                .store_trie_updates(block.block_with_parent(), BlockStateDiff::default())
+                .expect("store trie update");
+        }
+    }
+
+    fn init_storage<S: BaseProofsStore>(storage: BaseProofsStorage<S>) {
+        let genesis_block = NumHash::new(0, b256(0x00));
+        storage
+            .set_earliest_block_number(genesis_block.number, genesis_block.hash)
+            .expect("set earliest");
+        storage
+            .store_trie_updates(
+                BlockWithParent::new(genesis_block.hash, genesis_block),
+                BlockStateDiff::default(),
+            )
+            .expect("store trie update");
+    }
+
+    // Initialize exex with config
+    fn build_test_exex<NodeT, Store>(
+        ctx: ExExContext<NodeT>,
+        storage: BaseProofsStorage<Store>,
+    ) -> BaseProofsExEx<NodeT, Store>
+    where
+        NodeT: FullNodeComponents,
+        Store: BaseProofsStore + Clone + 'static,
+    {
+        BaseProofsExEx::builder(ctx, storage)
+            .with_proofs_history_window(20)
+            .with_proofs_history_prune_interval(Duration::from_secs(3600))
+            .with_verification_interval(1000)
+            .with_max_prune_blocks_startup(1000)
+            .build()
+    }
+
+    #[tokio::test]
+    async fn handle_notification_chain_committed() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        init_storage(proofs.clone());
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+
+        let new_chain = Arc::new(mk_chain_with_updates(1, 1, None));
+        let notif = ExExNotification::ChainCommitted { new: new_chain };
+
+        let sync_target = SyncTarget::new();
+
+        exex.handle_notification(notif, &sync_target).expect("handle chain commit");
+
+        // Committed blocks are cached in sync target, not stored directly
+        assert!(sync_target.take(1).is_some(), "block 1 should be cached");
+        let state = sync_target.take_state().expect("should have pending state");
+        assert!(matches!(state, SyncTargetState::SyncUpTo { to: 1 }));
+    }
+
+    #[tokio::test]
+    async fn handle_notification_chain_committed_caches_already_stored_blocks() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        init_storage(proofs.clone());
+
+        // Pre-store blocks 1..5 so storage is at block 5
+        store_blocks(1, 5, &proofs);
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+
+        let sync_target = SyncTarget::new();
+
+        // Handle notification for block 5 which is already stored - still caches
+        let new_chain = Arc::new(mk_chain_with_updates(5, 5, Some(hash_for_num(10))));
+        let notif = ExExNotification::ChainCommitted { new: new_chain };
+        exex.handle_notification(notif, &sync_target).expect("handle chain commit");
+
+        // State is set (sync loop will see latest >= target and skip)
+        let state = sync_target.take_state().expect("should have pending state");
+        assert!(matches!(state, SyncTargetState::SyncUpTo { to: 5 }));
+
+        // Storage is unchanged (notification handler doesn't write to storage)
+        let latest = proofs.get_latest_block_number().expect("get latest block").expect("ok");
+        assert_eq!(latest.0, 5);
+        assert_eq!(latest.1, hash_for_num(5));
+    }
+
+    #[tokio::test]
+    async fn handle_notification_chain_reorged() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        init_storage(proofs.clone());
+        store_blocks(1, 10, &proofs);
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+
+        let sync_target = SyncTarget::new();
+
+        // Now the tip is 10, and we want to reorg from block 6..12
+        let old_chain = Arc::new(mk_chain_with_updates(6, 10, None));
+        let new_chain = Arc::new(mk_chain_with_updates(6, 12, None));
+
+        // Notification: chain reorged 6..12
+        let notif = ExExNotification::ChainReorged { new: new_chain, old: old_chain };
+
+        exex.handle_notification(notif, &sync_target).expect("handle chain re-orged");
+
+        // Should have RevertThenSync state
+        let state = sync_target.take_state().expect("should have pending state");
+        assert!(matches!(
+            state,
+            SyncTargetState::RevertThenSync { revert_to, sync_to: 12 }
+            if revert_to.block.number == 6
+        ));
+
+        // New chain blocks should be cached
+        for n in 6..=12 {
+            assert!(sync_target.take(n).is_some(), "block {n} should be cached");
+        }
+
+        // Storage unchanged (sync loop handles the actual revert)
+        let latest = proofs.get_latest_block_number().expect("get latest block").expect("ok").0;
+        assert_eq!(latest, 10);
+    }
+
+    #[tokio::test]
+    async fn handle_notification_chain_reorged_beyond_stored_blocks() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        init_storage(proofs.clone());
+        store_blocks(1, 10, &proofs);
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+
+        let sync_target = SyncTarget::new();
+
+        // Now the tip is 10, and we want to reorg from block 12..15
+        // Both chains share the same fork block (block 11)
+        let old_chain = Arc::new(mk_chain_with_updates(12, 15, None));
+        let new_chain = Arc::new(mk_chain_with_updates(12, 20, None));
+
+        // Notification: chain reorged 12..20
+        let notif = ExExNotification::ChainReorged { new: new_chain, old: old_chain };
+
+        exex.handle_notification(notif, &sync_target).expect("handle chain re-orged");
+
+        // State is set; sync loop will detect revert is beyond stored blocks
+        let state = sync_target.take_state().expect("should have pending state");
+        assert!(matches!(
+            state,
+            SyncTargetState::RevertThenSync { revert_to, sync_to: 20 }
+            if revert_to.block.number == 12
+        ));
+
+        // Storage unchanged
+        let latest = proofs.get_latest_block_number().expect("get latest block").expect("ok").0;
+        assert_eq!(latest, 10);
+    }
+
+    #[tokio::test]
+    async fn handle_notification_chain_reverted() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        init_storage(proofs.clone());
+        store_blocks(1, 10, &proofs);
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+
+        let sync_target = SyncTarget::new();
+
+        // Now the tip is 10, and we want to revert from block 9..10
+        let old_chain = Arc::new(mk_chain_with_updates(9, 10, None));
+
+        // Notification: chain reverted 9..10
+        let notif = ExExNotification::ChainReverted { old: old_chain };
+
+        exex.handle_notification(notif, &sync_target).expect("handle chain reverted");
+
+        // Should have Revert state
+        let state = sync_target.take_state().expect("should have pending state");
+        assert!(matches!(
+            state,
+            SyncTargetState::Revert { revert_to }
+            if revert_to.block.number == 9
+        ));
+
+        // Storage unchanged (sync loop handles the actual revert)
+        let latest = proofs.get_latest_block_number().expect("get latest block").expect("ok").0;
+        assert_eq!(latest, 10);
+    }
+
+    #[tokio::test]
+    async fn handle_notification_chain_reverted_beyond_stored_blocks() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        init_storage(proofs.clone());
+        store_blocks(1, 5, &proofs);
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+
+        let sync_target = SyncTarget::new();
+
+        // Now the tip is 5, and we want to revert from block 9..10
+        let old_chain = Arc::new(mk_chain_with_updates(9, 10, None));
+
+        // Notification: chain reverted 9..10
+        let notif = ExExNotification::ChainReverted { old: old_chain };
+
+        exex.handle_notification(notif, &sync_target).expect("handle chain reverted");
+
+        // State is set; sync loop will detect revert is beyond stored blocks
+        let state = sync_target.take_state().expect("should have pending state");
+        assert!(matches!(
+            state,
+            SyncTargetState::Revert { revert_to }
+            if revert_to.block.number == 9
+        ));
+
+        // Storage unchanged
+        let latest = proofs.get_latest_block_number().expect("get latest block").expect("ok").0;
+        assert_eq!(latest, 5);
+    }
+
+    #[tokio::test]
+    async fn ensure_initialized_errors_on_storage_not_initialized() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+        let _ = exex.ensure_initialized().expect_err("should return error");
+    }
+
+    #[tokio::test]
+    async fn ensure_initialized_errors_when_prune_exceeds_threshold() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        init_storage(proofs.clone());
+
+        for i in 1..1100 {
+            proofs
+                .store_trie_updates(
+                    BlockWithParent::new(
+                        hash_for_num(i - 1),
+                        BlockNumHash::new(i, hash_for_num(i)),
+                    ),
+                    BlockStateDiff::default(),
+                )
+                .expect("store trie update");
+        }
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+        let _ = exex.ensure_initialized().expect_err("should return error");
+    }
+
+    #[tokio::test]
+    async fn ensure_initialized_succeeds() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        init_storage(proofs.clone());
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+        exex.ensure_initialized().expect("should not return error");
+    }
+
+    #[tokio::test]
+    async fn handle_notification_schedules_async_on_gap() {
+        // RocksDB proofs storage
+        let dir = tempdir_path();
+        let store = Arc::new(RocksdbProofsStorage::new(dir.as_path()).expect("env"));
+        let proofs: BaseProofsStorage<Arc<RocksdbProofsStorage>> = Arc::clone(&store).into();
+
+        init_storage(proofs.clone());
+
+        let (ctx, _handle) =
+            reth_exex_test_utils::test_exex_context().await.expect("exex test context");
+
+        let exex = build_test_exex(ctx, proofs.clone());
+
+        // Notification: chain committed 5..10 (Blocks 1,2,3,4 are missing from storage)
+        let new_chain = Arc::new(mk_chain_with_updates(5, 10, None));
+        let notif = ExExNotification::ChainCommitted { new: new_chain };
+
+        let sync_target = SyncTarget::new();
+
+        // Process notification
+        exex.handle_notification(notif, &sync_target)
+            .expect("handle chain commit should return ok immediately");
+
+        // Verify the sync target state was set
+        let state = sync_target.take_state().expect("should have pending state");
+        assert!(
+            matches!(state, SyncTargetState::SyncUpTo { to: 10 }),
+            "Should have scheduled sync to block 10"
+        );
+
+        // Verify Main Thread did NOT process it
+        // Because we didn't spawn the actual worker thread in this test, storage should still be at
+        // 0. This proves the 'handle_notification' returned instantly without doing the
+        // heavy lifting.
+        let latest = proofs.get_latest_block_number().expect("get").expect("ok").0;
+        assert_eq!(latest, 0, "Main thread should not have processed the blocks synchronously");
+    }
+}
